requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
langgraph==0.0.64
langchain==0.0.350
google-generativeai==0.3.2
Pillow==10.1.0
httpx==0.25.2
asyncio-mqtt==0.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
aioredis==2.0.1
celery==5.3.4